Refactors Code w.r.t. Subdirs Template
