+++ SimPlatform/GpaLib/borderlayout.cpp
