+++ SimPlatform/GpaLib/gpalib.cpp
@@ -1,0 +1,6 @@
+#include "gpalib.h"
+
+
+GpaLib::GpaLib()
+{
+}
