+++ SimPlatform/GpaLib/gpalib.h
@@ -1,0 +1,12 @@
+#ifndef GPALIB_H
+#define GPALIB_H
+
+
+class GpaLib
+{
+
+public:
+    GpaLib();
+};
+
+#endif // GPALIB_H
