+++ SimPlatform/GpaLib/text.cpp
