+++ SimPlatform/GpaLib/text.h
