+++ SimPlatform/Prototype1/Prototype1.pro
@@ -1,0 +1,55 @@
+#-------------------------------------------------
+#
+# Project created by QtCreator 2019-04-17T10:51:18
+#
+#-------------------------------------------------
+
+QT       += core gui
+
+greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
+
+TARGET = Prototype1
+TEMPLATE = app
+
+# The following define makes your compiler emit warnings if you use
+# any feature of Qt which has been marked as deprecated (the exact warnings
+# depend on your compiler). Please consult the documentation of the
+# deprecated API in order to know how to port your code away from it.
+DEFINES += QT_DEPRECATED_WARNINGS
+
+# You can also make your code fail to compile if you use deprecated APIs.
+# In order to do so, uncomment the following line.
+# You can also select to disable deprecated APIs only up to a certain version of Qt.
+#DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
+
+CONFIG += c++11
+
+SOURCES += \
+          *.cpp
+
+HEADERS += \
+          *.h
+
+FORMS += \
+        *.ui
+
+RESOURCES += \
+            *.qrc
+
+# Default rules for deployment.
+qnx: target.path = /tmp/$${TARGET}/bin
+else: unix:!android: target.path = /opt/$${TARGET}/bin
+!isEmpty(target.path): INSTALLS += target
+
+win32:CONFIG(release, debug|release): LIBS += -L$$OUT_PWD/../GpaLib/release/ -lGpaLib
+else:win32:CONFIG(debug, debug|release): LIBS += -L$$OUT_PWD/../GpaLib/debug/ -lGpaLib
+else:unix: LIBS += -L$$OUT_PWD/../GpaLib/ -lGpaLib
+
+INCLUDEPATH += $$PWD/../GpaLib
+DEPENDPATH += $$PWD/../GpaLib
+
+win32-g++:CONFIG(release, debug|release): PRE_TARGETDEPS += $$OUT_PWD/../GpaLib/release/libGpaLib.a
+else:win32-g++:CONFIG(debug, debug|release): PRE_TARGETDEPS += $$OUT_PWD/../GpaLib/debug/libGpaLib.a
+else:win32:!win32-g++:CONFIG(release, debug|release): PRE_TARGETDEPS += $$OUT_PWD/../GpaLib/release/GpaLib.lib
+else:win32:!win32-g++:CONFIG(debug, debug|release): PRE_TARGETDEPS += $$OUT_PWD/../GpaLib/debug/GpaLib.lib
+else:unix: PRE_TARGETDEPS += $$OUT_PWD/../GpaLib/libGpaLib.a
