+++ SimPlatform/Prototype1/maincontainerview.cpp
@@ -1,0 +1,13 @@
+#include "maincontainerview.h"
+#include "ui_maincontainerview.h"
+#include "container.h"
+
+MainWindow::MainWindow(QWidget *parent) :
+        QMainWindow(parent),
+        ui(new Ui::MainWindow) {
+    ui->setupUi(this);
+}
+
+MainWindow::~MainWindow() {
+    delete ui;
+}
