+++ SimPlatform/Prototype1/resources.qrc
@@ -1,0 +1,5 @@
+<RCC>
+    <qresource prefix="/resources">
+        <file>icon/gpa_logo.png</file>
+    </qresource>
+</RCC>
